Add For Loop for 256 JCheckBoxes of buildGUI()
Using a for loop, instantiates 256 unchecked JCheckboxes that will be added to an ArrayList of JCheckboxes (checkBoxList) and to the JPanel (mainPanel). This shall serve as the GUI component of which to be checked by the user to create beats. The ArrayList shall serve as the basis of which the program will check how to create the beats.