Created buildTrackAndStart() method
Created the buildTrackAndStart() method which is responsible for the playing of music in our application upon the trigger of the start button. Along the commit are the creation of an array variable that will hold integers, and a block of code to delete the old track played prior the trigger (or none at all) and create a new and empty track on top of it.